+++ modeling-core/src/main/java/org/inbio/modeling/core/dao/impl/GrassDAOImpl.java
@@ -127,7 +127,7 @@ public class GrassDAOImpl extends BaseDAOImpl implements GrassDAO {
 	}
 
 	@Override
-	public void executeWeightedSum(String layerName1, Long weight1, String layerName2, Long weight2, Long suffix, String outputName) throws Exception{
+	public void executeWeightedSum(String layerName1, Double weight1, String layerName2, Double weight2, Long suffix, String outputName) throws Exception{
 		int result = 0;
 		List<String> commands = null;
 		StringBuilder stdout = null;
