+++ modeling-core/src/main/java/org/inbio/modeling/core/manager/GrassManager.java
@@ -29,7 +29,7 @@ public interface GrassManager{
 
 	public void convertLayer2Raster(String layerName, Long suffix, String column) throws Exception;
 
-	public void executeWeightedSum(String layerName1, Long weight1, String layerName2, Long weight2, Long suffix, String outputName) throws Exception;
+	public void executeWeightedSum(String layerName1, Double weight1, String layerName2, Double weight2, Long suffix, String outputName) throws Exception;
 
 	public void exportLayer2Image(Long suffix, String layerName ) throws Exception;
