+++ modeling-core/src/main/java/org/inbio/modeling/core/manager/impl/GrassManagerImpl.java
@@ -53,7 +53,7 @@ public class GrassManagerImpl implements GrassManager {
 
 
 	@Override
-	public void executeWeightedSum(String layerName1, Long weight1, String layerName2, Long weight2, Long suffix, String outputName) throws Exception {
+	public void executeWeightedSum(String layerName1, Double weight1, String layerName2, Double weight2, Long suffix, String outputName) throws Exception {
 		this.grassDAOImpl.executeWeightedSum(layerName1, weight1, layerName2, weight2, suffix, outputName);
 	}
